- added db seeder - Fixed edit for post without previous image - added gitignore for public/article_images

+++ app/Http/Controllers/AdminController.php
@@ -346,8 +346,10 @@ class AdminController extends Controller {
 
             if ($success) {
 
-                if (isset($request->article_id)) {
-
+                
+                
+                
+                if (isset($request->article_id) && ($request->article_image_previous!="")) {                                        
                     $oldFileName = $request->article_image_previous;
                     unlink($oldFileName);
                 }
